% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{clipping}
\alias{clipping}
\title{Creates bit-level (clipping representation) from a vector}
\usage{
clipping(x)
}
\arguments{
\item{x}{the numeric vector (time series)}
}
\value{
the integer vector of zeros and ones
}
\description{
The \code{clipping} computes bit-level (clipping representation) from a vector.
}
\examples{
clipping(rnorm(50))

}
\references{
Bagnall A, Ratanamahatana C, Keogh E, Lonardi S, Janacek G (2006)
A bit level representation for time series data mining with shape based similarity.
Data Mining and Knowledge Discovery 13(1):11-40
}
\seealso{
\code{\link[TSrepr]{trending}}
}
\author{
Peter Laurinec, <tsreprpackage@gmail.com>
}
