% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_based_repr.R
\name{coefComp}
\alias{coefComp}
\alias{lmCoef}
\alias{coefComp}
\alias{rlmCoef}
\alias{coefComp}
\alias{l1Coef}
\title{Functions for linear regression model coefficients computation}
\usage{
lmCoef(X, Y)

rlmCoef(X, Y)

l1Coef(X, Y)
}
\arguments{
\item{X}{model (design) matrix of independent variables}

\item{Y}{vector of dependent variable}
}
\value{
Numeric vector of regression coefficients
}
\description{
Computes regression coefficients from a model.
}
\examples{
lmCoef(matrix(rnorm(10), ncol = 2), rnorm(5))

rlmCoef(matrix(rnorm(10), ncol = 2), rnorm(5))

l1Coef(matrix(rnorm(10), ncol = 2), rnorm(5))

}
