% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{elec_load}
\alias{elec_load}
\title{2 weeks of electricity load data from 50 consumers.}
\format{A data frame with 50 rows and 672 variables.}
\source{
Anonymized.
}
\usage{
elec_load
}
\description{
A dataset containing the electricity consumption time series from 50 consumers of the length of 2 weeks.
Every day is 48 measurements (half-hourly data). Each row represents one consumers time series.
}
\keyword{datasets}
