% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fast_stat}
\alias{fast_stat}
\alias{maxC}
\alias{minC}
\alias{meanC}
\alias{sumC}
\alias{medianC}
\title{Fast statistic functions (helpers)}
\usage{
maxC(x)

minC(x)

meanC(x)

sumC(x)

medianC(x)
}
\arguments{
\item{x}{the numeric vector}
}
\value{
the numeric value
}
\description{
Fast statistic functions (helpers) for representations computation.
}
\examples{
maxC(rnorm(50))

minC(rnorm(50))

meanC(rnorm(50))

sumC(rnorm(50))

medianC(rnorm(50))

}
\author{
Peter Laurinec, <tsreprpackage@gmail.com>
}
