% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nondata_adaptive_repr.R
\name{fftinv}
\alias{fftinv}
\title{Inverse FFT}
\usage{
fftinv(x)
}
\arguments{
\item{x}{Numeric vector of Fourier coefficients}
}
\value{
Numeric vector
}
\description{
\code{fftinv} computes inverse FFT.
}
\examples{
fftinv(fft(rnorm(50))[1:10])

}
