% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{maape}
\alias{maape}
\title{MAAPE}
\usage{
maape(x, y)
}
\arguments{
\item{x}{the numeric vector of real values}

\item{y}{the numeric vector of forecasted values}
}
\value{
the numeric value in \%
}
\description{
the \code{maape} computes MAAPE (Mean Arctangent Absolute Percentage Error) of a forecast.
}
\examples{
maape(runif(50), runif(50))

}
\references{
Sungil Kim, Heeyoung Kim (2016)
A new metric of absolute percentage error for intermittent demand forecasts,
 International Journal of Forecasting 32(3):669-679
}
\author{
Peter Laurinec, <tsreprpackage@gmail.com>
}
