% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mae}
\alias{mae}
\title{MAE}
\usage{
mae(x, y)
}
\arguments{
\item{x}{Numeric vector of real values}

\item{y}{Numeric vector of predicted values}
}
\value{
numeric value
}
\description{
\code{mae} Computes MAE of a forecast.
}
\examples{
mae(runif(50), runif(50))

}
