% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mape}
\alias{mape}
\title{MAPE}
\usage{
mape(x, y)
}
\arguments{
\item{x}{Numeric vector of real values}

\item{y}{Numeric vector of predicted values}
}
\value{
numeric value
}
\description{
\code{mape} Computes MAPE of a forecast.
}
\examples{
mape(runif(50), runif(50))

}
