% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mase}
\alias{mase}
\title{MASE}
\usage{
mase(real, forecast, naive)
}
\arguments{
\item{real}{Numeric vector of real values}

\item{forecast}{Numeric vector of predicted values}

\item{naive}{Numeric vector of naive forecast}
}
\value{
numeric value
}
\description{
\code{mase} Computes MASE of a forecast.
}
\examples{
mase(rnorm(50), rnorm(50), rnorm(50))

}
