% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mdae}
\alias{mdae}
\title{MDAE}
\usage{
mdae(x, y)
}
\arguments{
\item{x}{Numeric vector of real values}

\item{y}{Numeric vector of predicted values}
}
\value{
numeric value
}
\description{
\code{mdae} Computes MdAE of a forecast.
}
\examples{
mdae(runif(50), runif(50))

}
