% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{movave}
\alias{movave}
\title{Simple Moving Average}
\usage{
movave(x, order)
}
\arguments{
\item{x}{Numeric vector}

\item{order}{- order of simple moving average}
}
\value{
Numeric vector of smoothed values
}
\description{
\code{movave} computes Simple Moving Average (SMA) from a time series.
}
\examples{
movave(rnorm(50), 4)

}
