% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mse}
\alias{mse}
\title{MSE}
\usage{
mse(x, y)
}
\arguments{
\item{x}{the numeric vector of real values}

\item{y}{the numeric vector of forecasted values}
}
\value{
the numeric value
}
\description{
The \code{mse} computes MSE (Mean Squared Error) of a forecast.
}
\examples{
mse(runif(50), runif(50))

}
\author{
Peter Laurinec, <tsreprpackage@gmail.com>
}
