% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizations_R.R
\name{norm_atan}
\alias{norm_atan}
\title{Arctangent normalisation}
\usage{
norm_atan(x)
}
\arguments{
\item{x}{the numeric vector (time series)}
}
\value{
the numeric vector of normalised values
}
\description{
The \code{norm_atan} normalises time series by Arctangent to max (-1,1) range.
}
\examples{
norm_atan(rnorm(50))

}
\seealso{
\code{\link[TSrepr]{norm_z}, \link[TSrepr]{norm_min_max}}
}
\author{
Peter Laurinec, <tsreprpackage@gmail.com>
}
