% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{norm_min_max}
\alias{norm_min_max}
\title{Min-Max normalisation}
\usage{
norm_min_max(x)
}
\arguments{
\item{x}{the numeric vector (time series)}
}
\value{
the numeric vector of normalised values
}
\description{
The \code{norm_min_max} normalises time series by min-max method.
}
\examples{
norm_min_max(rnorm(50))

}
\seealso{
\code{\link[TSrepr]{norm_z}}
}
\author{
Peter Laurinec, <tsreprpackage@gmail.com>
}
