% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{norm_min_max_params}
\alias{norm_min_max_params}
\title{Min-Max normalisation with parameters}
\usage{
norm_min_max_params(x, min, max)
}
\arguments{
\item{x}{the numeric vector (time series)}

\item{min}{the numeric value}

\item{max}{the numeric value}
}
\value{
the numeric vector of normalised values
}
\description{
The \code{norm_min_max_params} normalises time series by min-max method with defined parameters.
}
\examples{
norm_min_max_params(rnorm(50), 0, 1)

}
\seealso{
\code{\link[TSrepr]{norm_z_params}}
}
\author{
Peter Laurinec, <tsreprpackage@gmail.com>
}
