% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{norm_z}
\alias{norm_z}
\title{Z normalization}
\usage{
norm_z(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
numeric vector of normalized values
}
\description{
\code{norm_z} normalizes data by z-score.
}
\examples{
norm_z(runif(50))

}
\seealso{
\code{\link[TSrepr]{norm_min_max}}
}
