% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{norm_z_params}
\alias{norm_z_params}
\title{Z-score normalisation with parameters}
\usage{
norm_z_params(x, mean, sd)
}
\arguments{
\item{x}{the numeric vector (time series)}

\item{mean}{the numeric value}

\item{sd}{the numeric value - standard deviation}
}
\value{
the numeric vector of normalised values
}
\description{
The \code{norm_z_params} normalises time series by z-score with defined mean and standard deviation.
}
\examples{
norm_z_params(runif(50), 0.5, 1)

}
\seealso{
\code{\link[TSrepr]{norm_min_max_params}}
}
\author{
Peter Laurinec, <tsreprpackage@gmail.com>
}
