% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nondata_adaptive_repr.R
\name{repr_dft}
\alias{repr_dft}
\title{DFT representation by FFT}
\usage{
repr_dft(x, coef)
}
\arguments{
\item{x}{Numeric vector}

\item{coef}{Number of coefficients to extract}
}
\value{
Numeric vector of DFT coefficients
}
\description{
\code{repr_dft} computes DFT representation from a time series by FFT.
}
\examples{
repr_dft(rnorm(50), coef = 4)

}
\seealso{
\code{\link[Tsrepr]{repr_dwt}}
}
