% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_based_repr.R
\name{repr_exp}
\alias{repr_exp}
\title{Exponential smoothing seasonal coefficients as representation}
\usage{
repr_exp(x, freq, alpha = TRUE, gamma = TRUE)
}
\arguments{
\item{x}{Numeric vector}

\item{freq}{frequency of the time series}

\item{alpha}{default to TRUE (automatic determination of smoothing factor), or number between 0 to 1}

\item{gamma}{default to TRUE (automatic determination of seasonal smoothing factor), or number between 0 to 1}
}
\value{
Numeric vector of seasonal coefficients
}
\description{
\code{repr_exp} computes exponential smoothing seasonal coefficients.
}
\examples{
repr_exp(rnorm(96), freq = 24)

}
\seealso{
\code{\link[Tsrepr]{repr_lm}, \link[Tsrepr]{repr_gam}, \link[Tsrepr]{repr_seas_profile}}
}
