% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_based_repr.R
\name{repr_gam}
\alias{repr_gam}
\title{GAM regression coefficients as representation}
\usage{
repr_gam(x, freq = NULL, xreg = NULL)
}
\arguments{
\item{x}{Numeric vector}

\item{freq}{frequency of the time series. Can be vector of two frequencies (seasonalities) or just an integer of one frequency.}

\item{xreg}{numeric vector or data.frame with additional regressors}
}
\value{
Numeric vector of GAM regression coefficients
}
\description{
\code{repr_gam} computes seasonal GAM regression coefficients.
}
\examples{
repr_gam(rnorm(96), freq = 24)

}
\seealso{
\code{\link[Tsrepr]{repr_lm}, \link[Tsrepr]{repr_exp}}
}
