% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repr_matrix.R
\name{repr_matrix}
\alias{repr_matrix}
\title{Computation of matrix of representations from matrix of time series}
\usage{
repr_matrix(x, fun = NULL, args = NULL, normalize = TRUE,
  fun_norm = norm_z)
}
\arguments{
\item{x}{matrix, data.frame or data.table of time series, where time series are in rows of a table}

\item{fun}{function that computes representation}

\item{args}{list of additional (or required) parameters of fun (function that computes representation)}

\item{normalize}{normalize (scale) representations? (default TRUE)}

\item{fun_norm}{normalization function (default \code{norm_z})}
}
\value{
Numeric matrix of representations of time series
}
\description{
\code{repr_matrix} Computation of matrix of representations from matrix of time series
}
\examples{
# Create random matrix of time series
mat_ts <- matrix(rnorm(100), ncol = 10)
repr_matrix(mat_ts, fun = repr_paa, args = list(q = 5, func = meanC))

}
