% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{repr_paa}
\alias{repr_paa}
\title{PAA - Piecewise Aggregate Approximation}
\usage{
repr_paa(x, q, func)
}
\arguments{
\item{x}{the numeric vector (time series)}

\item{q}{the integer of the length of the "piece"}

\item{func}{the aggregation function. Can be meanC, medianC, sumC, minC or maxC or similar aggregation function}
}
\value{
the numeric vector
}
\description{
The \code{repr_paa} computes PAA representation from a vector.
}
\examples{
repr_paa(rnorm(11), 2, meanC)

}
\seealso{
\code{\link[TSrepr]{repr_dwt}, \link[TSrepr]{repr_dft}}
}
