% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIP.R
\name{repr_pip}
\alias{repr_pip}
\title{PIP representation}
\usage{
repr_pip(x, times = 10)
}
\arguments{
\item{x}{the numeric vector (time series)}

\item{times}{the number of important points to extract (default 10)}
}
\value{
the integer vector of places of important points in a vector
}
\description{
The \code{repr_pip} computes PIP (Perceptually Important Points) representation from a time series.
}
\examples{
repr_pip(rnorm(100), times = 12)

}
