% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{repr_seas_profile}
\alias{repr_seas_profile}
\title{Mean seasonal profile of time series}
\usage{
repr_seas_profile(x, freq, func)
}
\arguments{
\item{x}{the numeric vector (time series)}

\item{freq}{the integer of the length of the season}

\item{func}{the aggregation function. Can be meanC or medianC or similar aggregation function.}
}
\value{
the numeric vector
}
\description{
The \code{repr_seas_profile} computes mean seasonal profile representation from a time series.
}
\examples{
repr_seas_profile(rnorm(48*10), 48, meanC)

}
\seealso{
\code{\link[TSrepr]{repr_lm}, \link[TSrepr]{repr_gam}, \link[TSrepr]{repr_exp}}
}
