% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{repr_sma}
\alias{repr_sma}
\title{Simple Moving Average representation}
\usage{
repr_sma(x, order)
}
\arguments{
\item{x}{the numeric vector (time series)}

\item{order}{the order of simple moving average}
}
\value{
the numeric vector of smoothed values of the length = length(x) - order + 1
}
\description{
The \code{repr_sma} computes Simple Moving Average (SMA) from a time series.
}
\examples{
repr_sma(rnorm(50), 4)

}
\author{
Peter Laurinec, <tsreprpackage@gmail.com>
}
