% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rleC}
\alias{rleC}
\title{RLE (Run Length Encoding) written in C++}
\usage{
rleC(x)
}
\arguments{
\item{x}{the integer vector (from \code{clipping} or \code{trending})}
}
\value{
the list of values and counts of zeros and ones
}
\description{
The \code{rleC} computes RLE from bit-level (clipping or trending representation) vector.
}
\examples{
# clipping
clipped <- clipping(rnorm(50))
rleC(clipped)
# trending
trended <- trending(rnorm(50))
rleC(trended)

}
