% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rmse}
\alias{rmse}
\title{RMSE}
\usage{
rmse(x, y)
}
\arguments{
\item{x}{Numeric vector of real values}

\item{y}{Numeric vector of predicted values}
}
\value{
numeric value
}
\description{
\code{rmse} Computes RMSE of a forecast.
}
\examples{
rmse(runif(50), runif(50))

}
