% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{smape}
\alias{smape}
\title{sMAPE}
\usage{
smape(x, y)
}
\arguments{
\item{x}{Numeric vector of real values}

\item{y}{Numeric vector of predicted values}
}
\value{
numeric value
}
\description{
\code{smape} Computes sMAPE of a forecast.
}
\examples{
smape(runif(50), runif(50))

}
