% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{synthetic_control}
\alias{synthetic_control}
\title{Simulated time series dataset.}
\format{A data frame with 400 rows and 61 variables.
Names of variables are \code{Class} (class number 1-6), \code{V2 - V61} (time series numeric values).}
\source{
\url{https://archive.ics.uci.edu/ml/datasets/Synthetic+Control+Chart+Time+Series}
}
\usage{
synthetic_control
}
\description{
This dataset contains 400 examples of control charts synthetically generated by the process in the paper Time-Series Similarity Queries Employing a Feature-Based Approach.
There are six different classes of control charts: 1. Normal 2. Cyclic 3. Increasing trend 4. Decreasing trend 5. Upward shift 6. Downward shift.
The first 200 rows of the data.frame (\code{synthetic_control[1:200,]}) are for training and
others are for testing (\code{synthetic_control[201:400,]}).
}
\keyword{datasets}
