% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{clipping}
\alias{clipping}
\title{Creates bit-level (clipping representation) from a vector}
\usage{
clipping(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
integer vector of zeros and ones
}
\description{
\code{clipping} Computes bit-level (clipping representation) from a vector.
}
\examples{
clipping(rnorm(50))

}
\seealso{
\code{\link[FeaClip]{trending}}
}
