% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mae}
\alias{mae}
\title{MAE}
\usage{
mae(x, y)
}
\arguments{
\item{x}{the numeric vector of real values}

\item{y}{the numeric vector of forecasted values}
}
\value{
the numeric value
}
\description{
The \code{mae} computes MAE (Mean Absolute Error) of a forecast.
}
\examples{
mae(runif(50), runif(50))

}
\author{
Peter Laurinec, <tsreprpackage@gmail.com>
}
