% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mape}
\alias{mape}
\title{MAPE}
\usage{
mape(x, y)
}
\arguments{
\item{x}{the numeric vector of real values}

\item{y}{the numeric vector of forecasted values}
}
\value{
the numeric value
}
\description{
the \code{mape} computes MAPE (Mean Absolute Percentage Error) of a forecast.
}
\examples{
mape(runif(50), runif(50))

}
\author{
Peter Laurinec, <tsreprpackage@gmail.com>
}
