% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mase}
\alias{mase}
\title{MASE}
\usage{
mase(real, forecast, naive)
}
\arguments{
\item{real}{the numeric vector of real values}

\item{forecast}{the numeric vector of forecasted values}

\item{naive}{the numeric vector of naive forecast}
}
\value{
the numeric value
}
\description{
The \code{mase} computes MASE (Mean Absolute Scaled Error) of a forecast.
}
\examples{
mase(rnorm(50), rnorm(50), rnorm(50))

}
\author{
Peter Laurinec, <tsreprpackage@gmail.com>
}
