% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mdae}
\alias{mdae}
\title{MdAE}
\usage{
mdae(x, y)
}
\arguments{
\item{x}{the numeric vector of real values}

\item{y}{the numeric vector of forecasted values}
}
\value{
the numeric value
}
\description{
The \code{mdae} computes MdAE (Median Absolute Error) of a forecast.
}
\examples{
mdae(runif(50), runif(50))

}
\author{
Peter Laurinec, <tsreprpackage@gmail.com>
}
