% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{norm_min_max}
\alias{norm_min_max}
\title{min-max normalization}
\usage{
norm_min_max(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
numeric vector of normalized values
}
\description{
\code{norm_min_max} normalizes data by min-max method.
}
\examples{
norm_min_max(rnorm(50))

}
\seealso{
\code{\link[FeaClip]{norm_z}}
}
