% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{norm_min_max_list}
\alias{norm_min_max_list}
\title{Min-Max normalization list}
\usage{
norm_min_max_list(x)
}
\arguments{
\item{x}{the numeric vector (time series)}
}
\value{
the list composed of:
 \describe{
 \item{\strong{norm_values}}{the numeric vector of normalised values of time series}
 \item{\strong{min}}{the min value}
 \item{\strong{max}}{the max value}
  }
}
\description{
The \code{norm_min_max_list} normalises time series by min-max method and returns normalization parameters (min and max).
}
\examples{
norm_min_max_list(rnorm(50))

}
\seealso{
\code{\link[TSrepr]{norm_z_list}}
}
