% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{norm_min_max_list}
\alias{norm_min_max_list}
\title{min-max normalization list}
\usage{
norm_min_max_list(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
list of numeric vector of normalized values and min-max values
}
\description{
\code{norm_min_max_list} normalizes data by min-max method and returns normalization parameters.
}
\examples{
norm_min_max_list(rnorm(50))

}
\seealso{
\code{\link[FeaClip]{norm_z_list}}
}
