% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{norm_z}
\alias{norm_z}
\title{Z-score normalisation}
\usage{
norm_z(x)
}
\arguments{
\item{x}{the numeric vector (time series)}
}
\value{
the numeric vector of normalised values
}
\description{
The \code{norm_z} normalises time series by z-score.
}
\examples{
norm_z(runif(50))

}
\seealso{
\code{\link[TSrepr]{norm_min_max}}
}
