% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{norm_z_list}
\alias{norm_z_list}
\title{Z normalization list}
\usage{
norm_z_list(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
list of numeric vector of normalized values and mean-sd values
}
\description{
\code{norm_z_list} normalizes data by z-score and returns normalization parameters.
}
\examples{
norm_z_list(runif(50))

}
\seealso{
\code{\link[FeaClip]{norm_min_max_list}}
}
