% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{norm_z_list}
\alias{norm_z_list}
\title{Z-score normalization list}
\usage{
norm_z_list(x)
}
\arguments{
\item{x}{the numeric vector (time series)}
}
\value{
the list composed of:
 \describe{
 \item{\strong{norm_values}}{the numeric vector of normalised values of time series}
 \item{\strong{mean}}{the mean value}
 \item{\strong{sd}}{the standard deviation}
  }
}
\description{
The \code{norm_z_list} normalizes time series by z-score and returns normalization parameters (mean and standard deviation).
}
\examples{
norm_z_list(runif(50))

}
\seealso{
\code{\link[TSrepr]{norm_min_max_list}}
}
\author{
Peter Laurinec, <tsreprpackage@gmail.com>
}
