% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nondata_adaptive_repr.R
\name{repr_dct}
\alias{repr_dct}
\title{DCT representation}
\usage{
repr_dct(x, coef)
}
\arguments{
\item{x}{Numeric vector}

\item{coef}{Number of coefficients to extract}
}
\value{
Numeric vector of DCT coefficients
}
\description{
\code{repr_dct} computes DCT representation from a time series.
}
\examples{
repr_dct(rnorm(50), coef = 4)

}
\seealso{
\code{\link[TSrepr]{repr_dft}, \link[TSrepr]{repr_dwt}}
}
