% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nondata_adaptive_repr.R
\name{repr_dwt}
\alias{repr_dwt}
\title{DWT representation}
\usage{
repr_dwt(x, level = 4, filter = "haar")
}
\arguments{
\item{x}{the numeric vector (time series)}

\item{level}{the level of DWT transformation (default is 4)}

\item{filter}{the filter name (default is "haar")}
}
\value{
the numeric vector of DWT coefficients
}
\description{
The \code{repr_dwt} computes DWT (Discrete Wavelet Transform) representation (coefficients) from a time series.
}
\examples{
repr_dwt(rnorm(50), level = 4)

}
\seealso{
\code{\link[TSrepr]{repr_dft}, \link[TSrepr]{repr_dct}}
}
