% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nondata_adaptive_repr.R
\name{repr_dwt}
\alias{repr_dwt}
\title{DWT representation}
\usage{
repr_dwt(x, level = 4, filter = "haar")
}
\arguments{
\item{x}{Numeric vector}

\item{level}{level of DWT transformation (default is 4)}

\item{filter}{filter name (default is "haar")}
}
\value{
Numeric vector of DWT coefficients
}
\description{
\code{repr_dwt} computes DWT representation (coefficients) from a time series.
}
\examples{
repr_dwt(rnorm(50), level = 4)

}
\seealso{
\code{\link[Tsrepr]{repr_dft}}
}
