% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_based_repr.R
\name{repr_exp}
\alias{repr_exp}
\title{Exponential smoothing seasonal coefficients as representation}
\usage{
repr_exp(x, freq, alpha = TRUE, gamma = TRUE)
}
\arguments{
\item{x}{the numeric vector (time series)}

\item{freq}{the frequency of the time series}

\item{alpha}{the smoothing factor (default to TRUE - automatic determination of smoothing factor), or number between 0 to 1}

\item{gamma}{the seasonal smoothing factor (default to TRUE - automatic determination of seasonal smoothing factor), or number between 0 to 1}
}
\value{
the numeric vector of seasonal coefficients
}
\description{
The \code{repr_exp} computes exponential smoothing seasonal coefficients.
}
\examples{
repr_exp(rnorm(96), freq = 24)

}
\seealso{
\code{\link[TSrepr]{repr_lm}, \link[TSrepr]{repr_gam}, \link[TSrepr]{repr_seas_profile},
\link[stats]{HoltWinters}}
}
