% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{repr_feaclip}
\alias{repr_feaclip}
\title{FeaClip representation of time series}
\usage{
repr_feaclip(x)
}
\arguments{
\item{x}{the numeric vector (time series)}
}
\value{
the numeric vector of length 8
}
\description{
The \code{repr_feaclip} computes representation of time series based on feature extraction from bit-level (clipping) representation.
}
\examples{
repr_feaclip(rnorm(50))

}
\seealso{
\code{\link[TSrepr]{repr_featrend}, \link[TSrepr]{repr_feacliptrend}}
}
\author{
Peter Laurinec, <tsreprpackage@gmail.com>
}
