% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{repr_feacliptrend}
\alias{repr_feacliptrend}
\title{FeaClipTrend representation of time series}
\usage{
repr_feacliptrend(x, func, pieces = 2L, order = 4L)
}
\arguments{
\item{x}{Numeric vector}

\item{func}{aggregation function for FeaTrend procedure}

\item{pieces}{number of parts of time series to split}

\item{order}{order of simple moving average}
}
\value{
numeric vector of frequences of features
}
\description{
\code{repr_feacliptrend} computes representation of time series based on feature extraction from bit-level and trend-level representation.
}
\examples{
repr_feacliptrend(rnorm(50), maxC, 2, 4)

}
\seealso{
\code{\link[FeaClip]{repr_featrend}, \link[FeaClip]{repr_feaclip}}
}
