% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{repr_featrend}
\alias{repr_featrend}
\title{FeaTrend representation of time series}
\usage{
repr_featrend(x, func, pieces = 2L, order = 4L)
}
\arguments{
\item{x}{Numeric vector}

\item{func}{function of aggregation, can be sumC or maxC or similar aggregation function}

\item{pieces}{number of parts of time series to split}

\item{order}{order of simple moving average}
}
\value{
numeric vector of length pieces
}
\description{
\code{repr_featrend} computes representation of time series based on feature extraction from trend-level representation.
}
\examples{
repr_featrend(rnorm(50), maxC)

}
\seealso{
\code{\link[TSrepr]{repr_feaclip}, \link[TSrepr]{repr_feacliptrend}}
}
