% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_based_repr.R
\name{repr_lm}
\alias{repr_lm}
\title{Linear model regression coefficients as representation}
\usage{
repr_lm(x, freq = NULL, method = "lm", xreg = NULL)
}
\arguments{
\item{x}{Numeric vector}

\item{freq}{frequency of the time series. Can be vector of two frequencies (seasonalities) or just an integer of one frequency.}

\item{method}{Linear regression method to use. It can be "lm", "rlm" or "l1".}

\item{xreg}{data.frame with additional regressors}
}
\value{
Numeric vector of regression coefficients
}
\description{
\code{repr_lm} computes seasonal regression coefficients.
}
\examples{
repr_lm(rnorm(96), freq = 24, method = "lm")

}
\seealso{
\code{\link[Tsrepr]{repr_gam}, \link[Tsrepr]{repr_exp}}
}
