% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repr_matrix.R
\name{repr_matrix}
\alias{repr_matrix}
\title{Computation of matrix of representations from matrix of time series}
\usage{
repr_matrix(x, func = NULL, args = NULL, normalize = TRUE,
  func_norm = norm_z)
}
\arguments{
\item{x}{the matrix, data.frame or data.table of time series, where time series are in rows of the table}

\item{func}{the function that computes representation}

\item{args}{the list of additional (or required) parameters of func (function that computes representation)}

\item{normalize}{normalise (scale) representations? (default is TRUE)}

\item{func_norm}{the normalization function (default is \code{norm_z})}
}
\value{
the numeric matrix of representations of time series
}
\description{
The \code{repr_matrix} computes matrix of representations from matrix of time series
}
\details{
TODO.
}
\examples{
# Create random matrix of time series
mat_ts <- matrix(rnorm(100), ncol = 10)
repr_matrix(mat_ts, func = repr_paa, args = list(q = 5, func = meanC))

}
