% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{repr_paa}
\alias{repr_paa}
\title{PAA - Piecewise Aggregate Approximation}
\usage{
repr_paa(x, q, func)
}
\arguments{
\item{x}{Numeric vector of real values}

\item{q}{integer of length of the "piece"}

\item{func}{aggregation function. Can be meanC, medianC, sumC, minC or maxC or similar aggregation function.}
}
\value{
numeric vector
}
\description{
\code{repr_paa} Computes PAA representation from a vector.
}
\examples{
repr_paa(rnorm(11), 2, meanC)

}
\seealso{
\code{\link[FeaClip]{repr_dwt}, \link[FeaClip]{repr_dft}}
}
