% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{repr_seas_profile}
\alias{repr_seas_profile}
\title{Seasonal profile of time series}
\usage{
repr_seas_profile(x, freq, func)
}
\arguments{
\item{x}{Numeric vector of real values}

\item{freq}{integer of length of the season}

\item{func}{aggregation function. Can be meanC or medianC or similar aggregation function.}
}
\value{
numeric vector
}
\description{
\code{repr_seas_profile} Computes Seasonal Profile representation from a vector.
}
\examples{
repr_seas_profile(rnorm(48*10), 48, meanC)

}
\seealso{
\code{\link[FeaClip]{repr_lm}, \link[FeaClip]{repr_gam}, \link[FeaClip]{repr_exp}}
}
