% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rleC}
\alias{rleC}
\title{RLE (Run Length Encoding) written in C++}
\usage{
rleC(x)
}
\arguments{
\item{x}{Numeric vector (from \code{clipping})}
}
\value{
list of values and counts of zeros and ones
}
\description{
\code{rleC} Computes RLE from bit-level (clipping representation) vector
}
\examples{
clipped <- clipping(rnorm(50))
rleC(clipped)

}
