% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rmse}
\alias{rmse}
\title{RMSE}
\usage{
rmse(x, y)
}
\arguments{
\item{x}{the numeric vector of real values}

\item{y}{the numeric vector of forecasted values}
}
\value{
the numeric value
}
\description{
The \code{rmse} computes RMSE (Root Mean Squared Error) of a forecast.
}
\examples{
rmse(runif(50), runif(50))

}
