% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{smape}
\alias{smape}
\title{sMAPE}
\usage{
smape(x, y)
}
\arguments{
\item{x}{the numeric vector of real values}

\item{y}{the numeric vector of forecasted values}
}
\value{
the numeric value in \%
}
\description{
The \code{smape} computes sMAPE (Symmetric Mean Absolute Percentage Error) of a forecast.
}
\examples{
smape(runif(50), runif(50))

}
\author{
Peter Laurinec, <tsreprpackage@gmail.com>
}
