% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{trending}
\alias{trending}
\title{Creates trend-level (trending) representation from a vector}
\usage{
trending(x)
}
\arguments{
\item{x}{the numeric vector (time series)}
}
\value{
the integer vector of zeros and ones
}
\description{
The \code{trending} Computes trend-level (trending) representation from a vector.
}
\examples{
trending(rnorm(50))

}
\seealso{
\code{\link[TSrepr]{clipping}}
}
\author{
Peter Laurinec, <tsreprpackage@gmail.com>
}
