% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{trending}
\alias{trending}
\title{Creates trend-level (bit-level) representation from a vector}
\usage{
trending(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
integer vector of zeros and ones
}
\description{
\code{trending} Computes trend-level (bit-level) representation from a vector.
}
\examples{
trending(rnorm(50))

}
\seealso{
\code{\link[TSrepr]{clipping}}
}
